- update for entity names - refactored package locations for endpoints
