+++ src/test/java/com/example/backendPlayground/unitTests/UserServiceTests.java
@@ -2,10 +2,10 @@ package com.example.backendPlayground.unitTests;
 
 import com.example.backendPlayground.exceptions.InvalidRequestDataException;
 import com.example.backendPlayground.exceptions.ResourceNotFoundException;
-import com.example.backendPlayground.user.User;
-import com.example.backendPlayground.user.UserDTO;
-import com.example.backendPlayground.user.UserRepository;
-import com.example.backendPlayground.user.UserService;
+import com.example.backendPlayground.features.user.User;
+import com.example.backendPlayground.features.user.UserDTO;
+import com.example.backendPlayground.features.user.UserRepository;
+import com.example.backendPlayground.features.user.UserService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
